update domain +add FinallCall +add InscriptionFinal +add Relationships

+++ Código/DDSDGuarani.Domain/Entities/Address.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
 using System.Text;
 
 namespace DDSDGuarani.Domain.Entities
@@ -8,13 +9,14 @@ namespace DDSDGuarani.Domain.Entities
     public class Address
     {
         [Key]
-        public long IdAddress{ get; set; }
+        public int IdAddress{ get; set; }
         
         public String StreetAndNumber { get; set; }
         public String Location { get; set; }
         public String PostalCode { get; set; }
         public String City { get; set; }
         public String Country { get; set; }
+        
         public User User { get; set; }
     }
 }
